removed source for guest image
This solves the issue of 404 not found for guest image. There is never a time when the avatar component needs a src for an image, as we are relying on the default. It throws an error because the src was not a path for an image, but instead just a string. Removing the string does not change the functionality, it just removes the error.

I also changed the string in case the user does not have an image loaded, but this is unlikely to happen at this stage, as the github default image should be present for the user.

Made types explicit for profile variables.