updated form labels
corrected the issues listed in the console by adding ids/HTMLFor to the violating nodes.

It seems like chakra UI has difficulties with both the date library we use, and its own checkboxes. The layout of the checkboxes is recommended by chakra UI on their component page, and I believe they should automatically be labelled by the FormLabel, but adding it manually has solved the problem.